refactor: util 함수들 캡슐화
